+++ Normalizer/ProductNormalizerInterface.php
@@ -21,6 +21,7 @@ interface ProductNormalizerInterface
      * @param array            $magentoAttributesOptions Attribute options list from Magento
      * @param string           $localeCode               The locale to apply
      * @param string           $scopeCode                The akeno scope
+     * @param array            $rootCategoryMapping      Root category mapping
      * @param boolean          $onlyLocalized            If true, only get translatable attributes
      *
      * @return array Computed data
@@ -31,6 +32,7 @@ interface ProductNormalizerInterface
         $magentoAttributesOptions,
         $localeCode,
         $scopeCode,
+        $rootCategoryMapping,
         $onlyLocalized = false
     );